Add support for 64-bit only builds
As a side-effect, this change removes the dependency on
build/target/board/*/BoardConfig.mk as there is no 64-bit only config to
use. Those BoardConfig files weren't really supposed to be inherited
(according to comments in them), so removing the dependency is probably a
good thing. As a result, several variables need to be set to defaults.

Signed-off-by: Rob Herring <robh@kernel.org>