Make s/w and h/w rendering support coexist
Make the DRM support init conditional on ro.kernel.qemu property. This
allows building both h/w and s/w rendering support into a single image and
selecting s/w rendering with "qemu=1" on the kernel command line. The
default is h/w rendering.

Ideally we would query the kernel to see if h/w rendering is available, but
that would need a script to run.

Signed-off-by: Rob Herring <robh@kernel.org>