Add cpuset initialization
The default cpuset initialization constrains everything to CPU0 which is
not desirable for multi-core systems. Most platforms hardcode the settings,
but we need to be able to work with a variable number of CPUs. So add an
init script to set the number of CPUs which can be used to initialize the
cpusets.

This setup is sufficient for homogeneous cores, but not big.LITTLE systems.
Once we have standard core capacity information, the setup can be adjusted.

Signed-off-by: Rob Herring <robh@kernel.org>