kconfig: decouple the defconfig and build output names
Currently, the defconfigs and build directories are 1-1. This means that no
2 defconfigs can share a build even if they just have trivial differences
like boot image parameters. Rework the lunch setup to decouple these
allowing multiple defconfigs to point to the same target build.

The relationship of names is complicated, so this is an attempt to document
them. It may not be completely accurate:
TARGET_PRODUCT is based on the lunch name which in turn is based on
defconfig files. AndroidProducts.mk needs to have a match on
TARGET_PRODUCT. The makefile AndroidProducts.mk points to must set
PRODUCT_DEVICE. TARGET_DEVICE is set from PRODUCT_DEVICE and are the same
name. The makefile listed in AndroidProducts.mk and BoardConfig.mk need
to live in a device directory matching TARGET_DEVICE. TARGET_DEVICE also
sets the build directory.

PRODUCT_DEVICE is set from kconfig CONFIG_PRODUCT_DEVICE if set. Otherwise,
the default is "linaro_$(TARGET_ARCH)".

Signed-off-by: Rob Herring <robh@kernel.org>