BoardConfig.mk: don't use LOCAL_PATH in included files
The LOCAL_PATH variable defined within the top-level BoardConfig.mk
for each included BoardConfig.mk doesn't seem to work properly. The
outcome of this issue is that SELinux labels aren't applied properly:

    $ ls -ltrZ /system/bin/conn_init
    -rwxr-xr-x 1 root shell u:object_r:system_file:s0 22740 2017-07-12 16:54 /system/bin/conn_init

With the patch applied, we do get the files correctly labeled:

    $ ls -ltrZ /system/bin/conn_init
    -rwxr-xr-x 1 root shell u:object_r:conn_init_exec:s0 22740 2017-07-12 16:54 /system/bin/conn_init

Also, simplify the inclusion of the BoardConfig.mk files in the parent
one.

Signed-off-by: Aleksander Morgado <aleksander@aleksander.es>