lowmem: Move lowmem.rc to vendor path
As part of Treble project in Android O, all the device specific files have
to be located in a separate vendor partition. Use TARGET_COPY_OUT_VENDOR
which will be set correctly based on split vendor partition or not.

Phony packages don't work for /vendor so just copy the file instead.

Signed-off-by: Rob Herring <robh@kernel.org>